save PO Number to database for authnet cc orders
Even though the po_number data is present in the REST payload, it is somehow getting stripped out before payment information is saved for the anet_creditcard payment method. So I added this after plugin which loads the just-saved order and adds the PO number from the $paymentMethod argument of afterSavePaymentInformationAndPlaceOrder, where it is present, then re-saves the order.