add poNumber to anet_creditcard config
This adds the 'poNumber' key to these configurations defined in vendor/authorizenet/magento-module-creditcard/etc/config.xml:

-paymentInfoKeys
-paymentAdditionalInformationKeys

This allows poNumber data to be added to the paymentMethod's additionalData array and included in order confirmation emails. The paymentInfoKeys configuration is referenced in vendor/magento/module-payment/Block/ConfigurableInfo.php:56 and determines which info is included when a payment's HTML is rendered. paymentAdditionalInformationKeys is referenced at \AuthorizeNet\Core\Observer\PaymentAdditionalDataAssignObserver::getAdditionalInformationList and only allows configured keys' values to be added to the additional_data array.